Figured out how to make the Side Menu Buttons load another view controller! Small things I  guess
