+++ Project_2/FirstViewController.swift
@@ -1,0 +1,29 @@
+//
+//  FirstViewController.swift
+//  Project_2
+//
+//  Created by Lam Nguyen on 5/16/21.
+//
+
+import UIKit
+
+class FirstViewController: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+
+        // Do any additional setup after loading the view.
+    }
+    
+
+    /*
+    // MARK: - Navigation
+
+    // In a storyboard-based application, you will often want to do a little preparation before navigation
+    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
+        // Get the new view controller using segue.destination.
+        // Pass the selected object to the new view controller.
+    }
+    */
+
+}
