+++ Project_2/UserViewController.swift
@@ -8,6 +8,7 @@
 import UIKit
 import SideMenu
 
+
 class UserViewController: UIViewController,  UIPickerViewDelegate, UIPickerViewDataSource {
     
     var menu: SideMenuNavigationController?
@@ -91,33 +92,4 @@ class UserViewController: UIViewController,  UIPickerViewDelegate, UIPickerViewD
 
 }
 
-class MenuListController: UITableViewController {
-    
-    var items = ["First","Second","Third","First","Second","Third"]
-    
-    let darkColor = UIColor(red: 33/255.0, green: 33/255.0, blue: 33/255.0, alpha: 1)
-    
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        tableView.backgroundColor = darkColor
-        tableView.register(UITableViewCell.self, forCellReuseIdentifier: "cell")
-    }
-    
-    override func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        return items.count
-    }
-    
-    override func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        let cell = tableView.dequeueReusableCell(withIdentifier: "cell", for: indexPath)
-        cell.textLabel?.text = items[indexPath.row]
-        cell.textLabel?.textColor = .white
-        cell.backgroundColor = darkColor
-        
-        return cell
-    }
-    
-    override func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
-        // Do something
-    }
-    
-}
+
